+++ beer_community/src/main/frontend/src/components/beerBox/beerBox.module.css
@@ -14,7 +14,7 @@
 }
 
 .img {
-    width: 33%;
+    width: 15%;
 }
 
 .text {
