+++ beer_community/src/main/frontend/src/components/beerList/beerList.jsx
@@ -4,14 +4,16 @@ import styles from "./beerList.module.css";
 const BeerList = (props) => {
   return (
     <div className={styles.mainBox}>
-      {props.data.beer_list.map((i) => (
+      {Array.isArray(props.data) ?
+        props.data.map((i) => (
         <BeerBox
           key={i.id}
-          image={i.image_url}
-          title={i.beer_name}
+          image={i.imageUrl}
+          title={i.beerName}
           description={i.information}
         />
-      ))}
+      ))
+      : null}
     </div>
   );
 };
