+++ beer_community/src/main/resources/application.yml
@@ -2,14 +2,14 @@ spring:
   datasource:
     url: jdbc:mysql://localhost:3306/beer_community?useSSL=false&characterEncoding=UTF-8&serverTimezone=UTC
     username: root
-    password: seojo1919@
+    password: 1234
     driver-class-name: com.mysql.cj.jdbc.Driver
   main:
     allow-circular-references: true
 
   jpa:
     hibernate:
-      ddl-auto: create
+      ddl-auto: none
     properties:
       hibernate:
         show_sql: false
